# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LUMEN Knowledge Base API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Users Endpoint",
        "operationId": "list_users_endpoint_api_admin_users_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Role"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_AdminUserView__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update User Endpoint",
        "operationId": "update_user_endpoint_api_admin_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_AdminUserView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/spaces": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List User Spaces Endpoint",
        "description": "查询用户已加入空间 + 可授予空间（API-054，REQ-050，维护态批5）。仅全局 admin。",
        "operationId": "list_user_spaces_endpoint_api_admin_users__user_id__spaces_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_AdminUserSpacesView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/import": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import File Endpoint",
        "operationId": "import_file_endpoint_api_import_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_file_endpoint_api_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_ImportFileView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/batch": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Batch Endpoint",
        "operationId": "import_batch_endpoint_api_import_batch_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_batch_endpoint_api_import_batch_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_BatchImportView_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/doc-links": {
      "get": {
        "tags": [
          "doc-links"
        ],
        "summary": "List Links Endpoint",
        "operationId": "list_links_endpoint_api_doc_links_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Id"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "outbound",
              "title": "Direction"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_list_DocLinkView__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "doc-links"
        ],
        "summary": "Create Link Endpoint",
        "operationId": "create_link_endpoint_api_doc_links_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocLinkCreateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope_DocLinkCreated_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminJoinedSpaceView": {
        "properties": {
          "space_id": {
            "type": "integer",
            "title": "Space Id"
          },
          "space_code": {
            "type": "string",
            "title": "Space Code"
          },
          "space_name": {
            "type": "string",
            "title": "Space Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "joined_at": {
            "type": "string",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "space_id",
          "space_code",
          "space_name",
          "role",
          "joined_at"
        ],
        "title": "AdminJoinedSpaceView"
      },
      "AdminUserSpacesView": {
        "properties": {
          "joined": {
            "items": {
              "$ref": "#/components/schemas/AdminJoinedSpaceView"
            },
            "type": "array",
            "title": "Joined"
          },
          "available": {
            "items": {
              "$ref": "#/components/schemas/AdminAvailableSpaceView"
            },
            "type": "array",
            "title": "Available"
          }
        },
        "type": "object",
        "required": [
          "joined",
          "available"
        ],
        "title": "AdminUserSpacesView"
      },
      "AdminUserView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_login_at": {
            "type": "string",
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "status",
          "last_login_at"
        ],
        "title": "AdminUserView"
      },
      "ApiEnvelope_AdminUserSpacesView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserSpacesView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[AdminUserSpacesView]"
      },
      "ApiEnvelope_AdminUserView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/AdminUserView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[AdminUserView]"
      },
      "ApiEnvelope_BatchImportView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/BatchImportView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[BatchImportView]"
      },
...
      "ApiEnvelope_DocLinkCreated_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/DocLinkCreated"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[DocLinkCreated]"
      },
...
      "ApiEnvelope_ImportFileView_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ImportFileView"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[ImportFileView]"
      },
...
      "ApiEnvelope_list_AdminUserView__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/AdminUserView"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[list[AdminUserView]]"
      },
      "ApiEnvelope_list_DocLinkView__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 0
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "default": "ok"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DocLinkView"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiEnvelope[list[DocLinkView]]"
      },
...
      "BatchImportView": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "success_count": {
            "type": "integer",
            "title": "Success Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "skipped_count": {
            "type": "integer",
            "title": "Skipped Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchImportItemView"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "total",
          "success_count",
          "failed_count",
          "skipped_count",
          "items"
        ],
        "title": "BatchImportView"
      },
      "Body_import_batch_endpoint_api_import_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "relative_paths": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relative Paths"
          },
          "conflict_policy": {
            "type": "string",
            "title": "Conflict Policy",
            "default": "skip"
          },
          "preserve_structure": {
            "type": "boolean",
            "title": "Preserve Structure",
            "default": true
          },
          "permission": {
            "$ref": "#/components/schemas/DocumentPermission",
            "default": "team"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_import_batch_endpoint_api_import_batch_post"
      },
      "Body_import_file_endpoint_api_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "permission": {
            "$ref": "#/components/schemas/DocumentPermission",
            "default": "team"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_file_endpoint_api_import_post"
      },
...
      "DocLinkCreateBody": {
        "properties": {
          "source_document_id": {
            "type": "integer",
            "title": "Source Document Id"
          },
          "link_text": {
            "type": "string",
            "title": "Link Text"
          },
          "target_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Document Id"
          },
          "target_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Title"
          },
          "link_type": {
            "type": "string",
            "title": "Link Type",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "source_document_id",
          "link_text"
        ],
        "title": "DocLinkCreateBody"
      },
      "DocLinkCreated": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "DocLinkCreated"
      },
      "DocLinkView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source_document_id": {
            "type": "integer",
            "title": "Source Document Id"
          },
          "target_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Document Id"
          },
          "target_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Title"
          },
          "link_text": {
            "type": "string",
            "title": "Link Text"
          },
          "link_type": {
            "type": "string",
            "title": "Link Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_document_id",
          "target_document_id",
          "target_title",
          "link_text",
          "link_type",
          "status"
        ],
        "title": "DocLinkView"
      },
...
      "DocumentPermission": {
        "type": "string",
        "enum": [
          "private",
          "team",
          "external"
        ],
        "title": "DocumentPermission",
        "description": "Document visibility levels defined by docs/06-db-design.md."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportFileView": {
        "properties": {
          "import_id": {
            "type": "integer",
            "title": "Import Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "parsed_doc_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed Doc Id"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          }
        },
        "type": "object",
        "required": [
          "import_id",
          "status",
          "parsed_doc_id",
          "chunk_count",
          "mode"
        ],
        "title": "ImportFileView"
      },
...
      "UpdateUserRequest": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "UpdateUserRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}